Fix up bottom nav
This makes the bottom nav match what we had before. Not sure why we had
to use content-sizing here (we just changed from rem to px, so I guess
that matters?). Probably would be better to use box-sizing everywhere,
but that would be a change in the main css, I think.