Start fleshing out Information Architecture
I've decided to try managing this content as pages in this repo (this
commit includes a markdown renderer for simplates), rather than managing
it as docstrings in simplates in the main repo. The reason is that the
full list of simplates is too fine-grained for our needs here. Here we
need a higher-level overview of the five basic pages on the site.