Scrap homepage. Simple TOC everywhere.
I was feeling really constrained by the homepage. This site needs to be
about the content! I actually looked at converting it to a static Jekyll
site, but I got mired trying to install nokogiri so I could use a TOC
plugin (and that would have been a pain to deploy anyway).