Yank contentious "To the Media" section; #17
After pushback from @patcon and further though, this section doesn't say
what I want it to say. It also is out of place on the "Brand" page. I'm
yanking it for now. The Interview Policy on my personal site is probably
the better place to say what I want to say here?