Write welcome doc
This adds a Welcome doc, and points to that from the homepage and from a
new notice at the top of every page. The purpose of the Welcome doc is
to articulate our open company philosophy, establishing different roles
in relation Gittip, each with their own expectations. The key subtext
here is that we don't expect users to play by our conflict resolution
rules, which is IMO the mistake I made that precipitated the long slow
decline into the Gittip crisis.