s/enable/cultivate in mission
With Gratipay 2.0 (https://github.com/gratipay/gratipay.com/issues/3419), we are taking a more proactive approach to our community. It's fitting, then, to change the verb in our mission from "enable," which we adopted because of its laissez-faire implications, to something stronger. I am choosing "cultivate" because I like the organic and agrarian overtones (as opposed to, e.g., "build" or "create" or "develop"). Thinking of our work as cultivation recognizes that Gratipay has a life of its own, and implies that we see ourselves as proactively yet carefully and sensitively shaping the direction this life takes in such a way that helps it to reach its fullest potential.