Just use merge commits in the security repo
Merge commits are easier to do in GitHub and they're our normal
practice. The only downside I see (specific to this situation--merge
commits have their detractors in general) is that we end up with a
commit comment in the main repo that points to a PR in the security
repo. Meh.